pyproject.toml
[project]
name = "ccupp"
version = "0.1.0"
description = "Chinese Common User Passwords Profiler — generate weak password dictionaries from personal information using social engineering insights"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pypinyin>=0.55.0",
    "pyyaml>=6.0",
    "rich>=14.2.0",
    "structlog>=25.5.0",
    "typer>=0.20.0",
]

[project.scripts]
ccupp = "ccupp.__main__:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
